feat: add alternative backend (#11)
* add the start of alternative backend

* docs: update readme for alternative backend

* feat(api): add system endpoint to fetch server settings

* fix(client): fix alternative backend modal

* fully support alternative backend in model management

* fix(api): fix model crash if alternative backend

* fix(client): not closing modal on model add and fix models refresh

* feat(api): add alternative backend support in chat part

* fix(api): non saving chat

* remove debug logs

* docs: add alternative backend example

* fix some code quality and move alternative backend to generation file

* fix(api): fix api crash on alternative backend

* docs: add mistral 7B instruct v0.2 example

* fix(api): alternative model can cause server crash

* fix(client): fix playground error message display

* docs: fix typo 7B instruct

* fix(clinet): remove alt backend model in non supported mode

* remove useless lines in examples

* feat: add api/client alt back authentication

* feat(api): handle alt backend authentication

* docs: update example to add alt backend auth compatibility

* fix(api): fix playground chat api

* fix(client) fix add model typo on alt backend

* fix(api): fix chat stop endpoint

+++ api/src/routes/models/index.ts
@@ -20,7 +20,9 @@ router.get('/', async (req, res) => {
   res.json(models.map((model) => ({
     id: model._id,
     name: model.name,
-    path: `${process.env.MODELS_DIR}/${model.path}`,
+    path: model.alternativeBackend ? model.path :
+      `${process.env.MODELS_DIR}/${model.path}`,
+    alternativeBackend: model.alternativeBackend,
     createdAt: model.createdAt
   })));
 });
@@ -35,6 +37,7 @@ router.get('/:id', async (req, res) => {
     name: model.name,
     createdAt: model.createdAt,
     parameters: model.parameters,
+    alternativeBackend: model.alternativeBackend,
     promptTemplate: model.chatPromptTemplate
   });
 });
@@ -44,7 +47,8 @@ router.delete('/:id', async (req, res) => {
   if (!model) {
     return res.status(404).json({ message: 'Model not found' });
   }
-  await file.delete(`${process.env.MODELS_DIR}/${model.path}`);
+  if (!model.alternativeBackend)
+    await file.delete(`${process.env.MODELS_DIR}/${model.path}`);
   await mongoose.model('Models').findByIdAndDelete(req.params.id);
   res.status(200).json({ message: 'Model deleted' });
 });
@@ -53,7 +57,9 @@ router.post('/', async (req, res) => {
   const schema = yup.object().shape({
     name: yup.string().required(),
     uri: yup.string().required(),
-    promptTemplate: yup.string().required()
+    alternativeBackend: yup.boolean().default(false),
+    promptTemplate: yup.string().optional(),
+    parameters: yup.object().default({}).optional()
   })
   let payload: yup.InferType<typeof schema>;
   try {
@@ -61,29 +67,46 @@ router.post('/', async (req, res) => {
   } catch (e) {
     return res.status(400).json({ message: 'Missing required fields' });
   }
-  try {
-    compileTemplate(payload.promptTemplate, { system: 'system', messages: [{role: 'user', message: ''}] });
-  } catch (e) {
-    return res.status(400).json({ message: 'Invalid prompt template' });
+  if (!payload.alternativeBackend) {
+    if (!payload.promptTemplate) {
+      return res.status(400).json({ message: 'Missing prompt template' });
+    }
+    try {
+      compileTemplate(payload.promptTemplate, { system: 'system', messages: [{role: 'user', message: ''}] });
+    } catch (e) {
+      return res.status(400).json({ message: 'Invalid prompt template' });
+    }
   }
   const ModelObj = mongoose.model('Models');
-  const fileName = encodeURI(payload.name);
-  file.download(payload.uri, `${process.env.MODELS_DIR}/${fileName}`)
-    .then(() => {
-      const model = new ModelObj({
-        name: payload.name,
-        path: `${fileName}`,
-        chatPromptTemplate: payload.promptTemplate,
-        parameters: req.body.parameters
+  if (!payload.alternativeBackend) {
+    const fileName = encodeURI(payload.name);
+    file.download(payload.uri, `${process.env.MODELS_DIR}/${fileName}`)
+      .then(() => {
+        const model = new ModelObj({
+          name: payload.name,
+          path: `${fileName}`,
+          chatPromptTemplate: payload.promptTemplate,
+          parameters: req.body.parameters
+        });
+        model.save();
       });
-      model.save();
-    })
     res.status(202).json({ message: 'Model download in progress.' });
+  } else {
+    const model = new ModelObj({
+      name: payload.name,
+      path: payload.uri,
+      alternativeBackend: true,
+      parameters: payload.parameters
+    });
+    model.save();
+    res.status(201).json({ message: 'Model created' });
+  }
 });
 
 router.patch('/:id', async (req, res) => {
   const schema = yup.object().shape({
-    promptTemplate: yup.string().required()
+    promptTemplate: yup.string().optional(),
+    parameters: yup.object().default({}).optional()
   })
   let payload: yup.InferType<typeof schema>
 
@@ -94,12 +117,16 @@ router.patch('/:id', async (req, res) => {
     return res.status(400).json({ message: 'Bad request' });
   }
   try {
-    compileTemplate(payload.promptTemplate, { system: 'system', messages: [{role: 'user', message: ''}] });
+    if (payload.promptTemplate)
+      compileTemplate(payload.promptTemplate, { system: 'system', messages: [{role: 'user', message: ''}] });
   } catch (e) {
     return res.status(400).json({ message: 'Invalid prompt template' });
   }
   try {
-    await mongoose.model('Models').findOneAndUpdate({ _id: req.params.id }, { chatPromptTemplate: payload.promptTemplate });
+    await mongoose.model('Models').findOneAndUpdate({ _id: req.params.id },
+      { chatPromptTemplate: payload.promptTemplate,
+        parameters: Object.keys(payload.parameters).length > 0 ? payload.parameters : undefined
+      });
   } catch (e) {
     console.error(e)
     return res.status(500).json({ message: 'Internal Server Error' })
