+++ api/src/routes/playground/chat.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response, NextFunction } from 'express';
-import { Generation, GenerationLaunchOutput } from '../../utils/generation';
+import { Generation, GenerationOutput } from '../../utils/generation';
 import * as yup from 'yup';
 import * as Sentry from '@sentry/node';
 import mongoose from 'mongoose';
@@ -28,7 +28,7 @@ router.post('/', async (req: Request, res: Response, next: NextFunction) => {
   let ignoreIndex = 0;
   let response = '';
   let detecting = false;
-  let child: GenerationLaunchOutput;
+  let child: GenerationOutput;
   const transaction = Sentry.getActiveTransaction();
   let span: Sentry.Span|undefined;
   if (!process.env.LLAMA_PATH || !process.env.MODELS_DIR) {
@@ -51,23 +51,27 @@ router.post('/', async (req: Request, res: Response, next: NextFunction) => {
       span.finish();
     return res.status(400).json({ message: 'Model not found' });
   }
-  prompt += compileTemplate(model.chatPromptTemplate, { system: payload.system, messages: payload.messages });
-  console.log(prompt);
+  if (!model.alternativeBackend)
+    prompt += compileTemplate(model.chatPromptTemplate, { system: payload.system, messages: payload.messages });
   if (span)
     span.finish();
   try {
     if (transaction)
       span = transaction.startChild({ op: 'generation', description: 'Generate response' });
-    child = generation.launch({
-      modelPath: `${process.env.MODELS_DIR}/${model.path}`,
-      contextSize: 4096,
-      temperature: 0.7,
-      repeatPenalty: 1.1,
-      threads: 4,
-      nPredict: -1,
-      prompt: prompt,
-      interactive: false,
-    });
+    if (!model.alternativeBackend) {
+      child = generation.launch({
+        modelPath: `${process.env.MODELS_DIR}/${model.path}`,
+        contextSize: 4096,
+        temperature: 0.7,
+        repeatPenalty: 1.1,
+        threads: 4,
+        nPredict: -1,
+        prompt: prompt,
+        interactive: false,
+      });
+    } else {
+      child = await generation.generateCompletionAlt(payload.messages, model.path, model.parameters.authentication);
+    }
   } catch (e) {
     return res.status(500).json({ message: 'Something went wrong' });
   }
@@ -80,7 +84,12 @@ router.post('/', async (req: Request, res: Response, next: NextFunction) => {
   res.status(200);
   res.flushHeaders();
 
-  child.stdout.on('data', (data) => {
+  child.data.on('data', (data) => {
+    if (model.alternativeBackend) {
+      res.write(data.toString());
+      res.flushHeaders();
+      return;
+    }
     if (ignoreIndex < prompt.replaceAll(/<\/?s>/g, '').trim().length) {
       ignoreIndex += data.toString().length;
       if (process.env.NODE_ENV === 'development')
@@ -93,14 +102,16 @@ router.post('/', async (req: Request, res: Response, next: NextFunction) => {
     }
   });
 
-  child.stderr.on('data', (data) => {
-    if (process.env.NODE_ENV === 'development')
-      console.error(`stderr: ${data}`);
-    if (data.toString().includes('[end of text]'))
-      child.kill();
-  });
+  if (!model.alternativeBackend) {
+    child.stderr!.on('data', (data) => {
+      if (process.env.NODE_ENV === 'development')
+        console.error(`stderr: ${data}`);
+      if (data.toString().includes('[end of text]'))
+        child.kill();
+    });
+  }
 
-  child.on('close', (code) => {
+  child.data.on('close', () => {
     if (span)
       span.finish();
     res.end();
